`Furniture`: Inline `Furniture` into `FurniturePlacement`
This gets us almost all the way to the rename; since we don't have the
`Furniture` namespace in the way...
